+++ lib/resource/color/ResourceColor.dart
@@ -2,23 +2,19 @@
 
 class ResourceColor {
 
-
-  static var black = "#31312e";
-  static var blueLight = "#89bdfd";
-
-  //دراجته
-  static var blueDark = "#32759d";
-  static var blueDark_25 = blueDark; //more darker
-
   static var white = "#ffffff";
-  static var blue_aquamarine = "#64a8ab";
+  static var black = "#000000";
+  static var gray_dark_blue  = "#999999";
   static var orange = "#ffb908";
 
-  static var blueGreen_transparent = "#9064a8ab";
+  static var blueLight = "#89bdfd";
+  static var blueDark = "#32759d";
+  static var blue_aquamarine = "#64a8ab";
 
+  static var blueGreen_transparent = "#9064a8ab";
   static var blueDark_transparent = "#50637cfd";
 
-  static var gray_dark_blue  = "#5d6c72";
+
 
 
 }