+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.133
+ 
+* edit readme for button
+
+## 1.0.132
+
+* Default button color to black/gray
+
+## 1.0.131
+
+* Update "Readme"
+
 ## 1.0.130
 
 * Fix "ScrollViewPage.dart" in webplatform for parameter "thumbVisibility"